requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.3
aiokafka==0.10.0
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
ariadne==0.22
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
asyncer==0.0.2
asyncio==3.4.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.0
binaryornot==0.4.4
black==24.2.0
blinker==1.7.0
boto3==1.34.54
botocore==1.34.54
Brotli==1.1.0
bs4==0.0.2
build==1.0.3
CacheControl==0.13.1
cachelib==0.12.0
cachetools==5.3.2
cattrs==23.2.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
chainlit==1.0.301
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.8.2
comm==0.2.1
contourpy==1.2.0
cookiecutter==2.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.3
crashtest==0.4.1
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
data-object==1.5.0
dataclasses-json==0.5.14
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.8.0rc2
Deprecated==1.2.14
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.2
django-admin-honeypot==1.1.0
django-anymail==10.2
django-appconf==1.0.6
django-autoslug==1.9.9
django-bootstrap4==24.1
django-bootstrap5==23.4
django-celery==3.1.17
django-celery-results==2.5.1
django-classy-tags==4.1.0
django-cors-headers==4.3.1
django-countries==7.5.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-imagekit==5.0.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-ninja==1.1.0
django-queued-storage==0.8.0
django-redis==5.4.0
django-redis-cache==3.0.1
django-render-block==0.9.2
django-rest-knox==4.2.0
django-rest-passwordreset==1.4.0
django-rest-swagger==2.2.0
django-sekizai==4.1.0
django-templated-email==3.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
dm-tree==0.1.8
dnspython==2.6.1
docker==7.0.0
drf-chunked-upload==0.6.0
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.4.1
drf-yasg==1.21.7
dulwich==0.21.7
eli5==0.13.0
entrypoints==0.4
eventlet==0.36.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.89.1
fastapi-socketio==0.0.10
fastjsonschema==2.19.1
filelock==3.13.1
filetype==1.2.0
flake8==7.0.0
Flask==3.0.2
Flask-Cache==0.13.1
Flask-Classy==0.6.10
Flask-DebugToolbar==0.14.1
Flask-Humanize==0.3.0
Flask-Migrate==4.0.5
Flask-OAuthlib==0.9.6
flask-shell-ipython==0.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.5.4
gevent==24.2.1
geventhttpclient==2.0.2
gitdb==4.0.11
GitPython==3.1.42
google-pasta==0.2.0
googleapis-common-protos==1.62.0
gql==3.5.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
graphviz==0.20.1
greenlet==3.0.3
grpcio==1.62.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
hiredis==2.3.2
html5tagger==1.3.0
httpcore==1.0.4
httpie==3.2.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
humanize==4.9.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.1.2
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
ipdb==0.13.13
ipykernel==6.29.3
ipython==8.22.2
ipyvue==1.10.2
ipyvuetify==1.9.1
ipywidgets==8.1.2
isort==5.13.2
itsdangerous==2.1.2
itypes==1.2.0
jaraco.classes==3.3.1
jedi==0.19.1
Jinja2==3.1.3
jinja2-time==0.2.0
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
keras==3.0.5
keras-nightly==3.1.0.dev2024030603
keyring==24.3.0
kiwisolver==1.4.5
kombu==5.3.5
Lazify==0.4.0
libclang==16.0.6
linecache2==1.0.0
literalai==0.0.204
lxml==5.1.0
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
mlflow==2.11.0
mlserver==1.3.5
mlserver-mlflow==1.3.5
mock==5.1.0
more-itertools==10.2.0
mpi4py==3.1.5
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
nameko==2.14.1
namex==0.0.7
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
oauth2==1.9.0.post1
oauthlib==2.1.0
openapi-codec==1.3.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opt-einsum==3.3.0
orjson==3.9.15
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
parso==0.8.3
path==16.12.0
path.py==12.5.0
pathspec==0.12.1
pexpect==4.9.0
pilkit==3.0
pillow==10.2.0
pip-tools==7.4.1
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.4.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
premailer==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg==3.1.18
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-grpc-prometheus==0.8.0
pyarrow==15.0.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.2
pydeck==0.8.1b0
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pymdown-extensions==10.7.1
PyMySQL==1.1.0
pyparsing==3.1.1
pypdf==4.1.0
PyPDF2==3.0.1
PyPDF3==1.0.6
pyproject_hooks==1.0.0
PySocks==1.7.1
pytest==8.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-memcached==1.62
python-multipart==0.0.6
python-rapidjson==1.16
python-slugify==8.0.4
python-socketio==5.11.1
python3-openid==3.2.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
querystring-parser==1.2.4
rapidfuzz==3.6.1
reacton==1.8.2
reactpy==1.0.2
redis==3.5.3
referencing==0.33.0
requests==2.31.0
requests-cache==1.2.0
requests-oauthlib==1.1.0
requests-toolbelt==1.0.0
rich==13.7.0
rich-click==1.7.3
rpds-py==0.18.0
Rx==3.2.0
s3transfer==0.10.0
sanic==23.12.1
sanic-routing==23.12.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
selenium==4.18.1
setuptools==69.1.1
shellingham==1.5.4
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
solara==1.28.0
sortedcontainers==2.4.0
soupsieve==2.5
spark==0.2.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.22.0
starlette-exporter==0.17.1
strawberry-graphql==0.219.2
streamlit==1.31.1
stripe==8.9.0
sympy==1.12
syncer==2.0.3
tabulate==0.9.0
tb-nightly==2.17.0a20240305
tenacity==8.2.3
tensorboard-data-server==0.7.2
termcolor==2.4.0
text-unidecode==1.3
tf-nightly==2.17.0.dev20240305
tf_nightly_intel==2.17.0.dev20240305
threadpoolctl==3.3.0
tinyml==0.0.6
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traceback2==1.4.0
tracerite==1.1.1
traitlets==5.14.1
trame==3.5.2
trame-client==2.16.2
trame-server==2.17.2
trame-vtk==2.8.5
trio==0.24.0
trio-websocket==0.11.1
tritonclient==2.43.0
trove-classifiers==2024.2.23
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
unittest2==1.1.0
uptrace==1.22.0
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.0.7
uvicorn==0.25.0
validators==0.22.0
vine==5.1.0
virtualenv==20.25.0
waitress==3.0.0
watchdog==4.0.0
watchfiles==0.20.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
wheel==0.42.0
whitenoise==6.6.0
widgetsnbextension==4.0.10
wrapt==1.16.0
wslink==1.12.4
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.17.0
zope.event==5.0
zope.interface==6.2
